# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank project",
    "description": "第一步：通用数据接入与暂存。第二步：按表进行本体对齐、展示映射及生成独立图谱版本。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/intake/limits": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Limits",
        "operationId": "limits_api_v1_intake_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Limits"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/intake/uploads": {
      "post": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Upload",
        "operationId": "upload_api_v1_intake_uploads_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Filename"
            }
          },
          {
            "name": "header_row",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 1,
              "title": "Header Row"
            }
          },
          {
            "name": "encoding",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "utf-8-sig",
                "gb18030"
              ],
              "type": "string",
              "default": "auto",
              "title": "Encoding"
            }
          },
          {
            "name": "delimiter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "comma",
                "tab",
                "semicolon"
              ],
              "type": "string",
              "default": "auto",
              "title": "Delimiter"
            }
          },
          {
            "name": "skip_description_sheets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Skip Description Sheets"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BatchDetail"
                    },
                    {
                      "$ref": "#/components/schemas/IntakeJob"
                    }
                  ],
                  "title": "Response Upload Api V1 Intake Uploads Post"
                }
              }
            }
          },
          "202": {
            "description": "文件已保存，等待后台解析",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/mysql/catalog": {
      "post": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Mysql Catalog",
        "operationId": "mysql_catalog_api_v1_intake_mysql_catalog_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MysqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogTable"
                  },
                  "type": "array",
                  "title": "Response Mysql Catalog Api V1 Intake Mysql Catalog Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/intake/mysql/import": {
      "post": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Mysql Import",
        "operationId": "mysql_import_api_v1_intake_mysql_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MysqlImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BatchDetail"
                    },
                    {
                      "$ref": "#/components/schemas/IntakeJob"
                    }
                  ],
                  "title": "Response Mysql Import Api V1 Intake Mysql Import Post"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/intake/batches": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Batches",
        "operationId": "batches_api_v1_intake_batches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/batches/{batch_id}": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Batch",
        "operationId": "batch_api_v1_intake_batches__batch_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Delete",
        "operationId": "delete_api_v1_intake_batches__batch_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/batches/{batch_id}/tables/{table_id}": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Table",
        "operationId": "table_api_v1_intake_batches__batch_id__tables__table_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Table Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/jobs": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Intake Jobs",
        "operationId": "intake_jobs_api_v1_intake_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntakeJob"
                  },
                  "type": "array",
                  "title": "Response Intake Jobs Api V1 Intake Jobs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/intake/jobs/{job_id}": {
      "get": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Intake Job",
        "operationId": "intake_job_api_v1_intake_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Retry Intake",
        "operationId": "retry_intake_api_v1_intake_jobs__job_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "第一步 · 数据接入"
        ],
        "summary": "Cancel Intake",
        "operationId": "cancel_intake_api_v1_intake_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_kind": {
            "type": "string",
            "enum": [
              "file",
              "mysql"
            ],
            "title": "Source Kind"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "table_count": {
            "type": "integer",
            "title": "Table Count"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableInfo"
            },
            "type": "array",
            "title": "Tables"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "source_kind",
          "source",
          "created_at",
          "table_count",
          "row_count",
          "tables"
        ],
        "title": "BatchDetail"
      },
...
      "BatchPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "BatchPage"
      },
      "CatalogTable": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "estimated_rows": {
            "type": "integer",
            "title": "Estimated Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "comment",
          "estimated_rows"
        ],
        "title": "CatalogTable"
      },
      "Column": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type",
            "default": "text"
          },
          "type_origin": {
            "type": "string",
            "enum": [
              "observed",
              "declared"
            ],
            "title": "Type Origin",
            "default": "observed"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable",
            "default": true
          },
          "comment": {
            "type": "string",
            "title": "Comment",
            "default": ""
          },
          "primary_key": {
            "type": "boolean",
            "title": "Primary Key",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Column"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "rows_done": {
            "type": "integer",
            "title": "Rows Done"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "status",
          "name",
          "size_bytes",
          "rows_done",
          "attempt",
          "batch_id",
          "error"
        ],
        "title": "IntakeJob"
      },
      "Limits": {
        "properties": {
          "max_upload_bytes": {
            "type": "integer",
            "title": "Max Upload Bytes",
            "default": 20971520
          },
          "max_expanded_bytes": {
            "type": "integer",
            "title": "Max Expanded Bytes",
            "default": 104857600
          },
          "max_tables": {
            "type": "integer",
            "title": "Max Tables",
            "default": 30
          },
          "max_rows": {
            "type": "integer",
            "title": "Max Rows",
            "default": 50000
          },
          "max_columns": {
            "type": "integer",
            "title": "Max Columns",
            "default": 512
          },
          "max_cells": {
            "type": "integer",
            "title": "Max Cells",
            "default": 2000000
          },
          "max_cell_chars": {
            "type": "integer",
            "title": "Max Cell Chars",
            "default": 100000
          }
        },
        "type": "object",
        "title": "Limits"
      },
...
      "MysqlConnection": {
        "properties": {
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "default": 3306
          },
          "database": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Database"
          },
          "user": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "User"
          },
          "password": {
            "type": "string",
            "maxLength": 4096,
            "format": "password",
            "title": "Password",
            "default": "",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "host",
          "database",
          "user"
        ],
        "title": "MysqlConnection"
      },
      "MysqlImport": {
        "properties": {
          "connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MysqlConnection"
              },
              {
                "type": "null"
              }
            ]
          },
          "tables": {
            "items": {
              "type": "string",
              "maxLength": 255,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Tables"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tables"
        ],
        "title": "MysqlImport"
      },
      "MysqlRequest": {
        "properties": {
          "connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MysqlConnection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MysqlRequest"
      },
...
      "TableInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "type": "array",
            "title": "Columns"
          },
          "foreign_keys": {
            "items": {
              "$ref": "#/components/schemas/ForeignKey"
            },
            "type": "array",
            "title": "Foreign Keys"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "row_count",
          "columns"
        ],
        "title": "TableInfo"
      },
...
      "TablePage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "type": "array",
            "title": "Columns"
          },
          "foreign_keys": {
            "items": {
              "$ref": "#/components/schemas/ForeignKey"
            },
            "type": "array",
            "title": "Foreign Keys"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/DataRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "row_count",
          "columns",
          "rows",
          "offset",
          "limit"
        ],
        "title": "TablePage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}